식당상세, 예약, 리뷰, 신고 연계 완성

+++ src/main/java/com/se_b4/catchtable/entity/ReportEntity.java
@@ -19,7 +19,7 @@ public class ReportEntity
     private int dining_uid;
 
     @NotNull
-    private int reporter_uuid;
+    private Long reporter_uuid;
 
     private boolean illegal_service;
     private boolean poor_hygiene;
